+++ log/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0](https://github.com/ankorstore/yokai/compare/log/v1.0.0...log/v1.1.0) (2024-02-22)
+
+
+### Features
+
+* **log:** Updated default logger and added possibility to dump logs test buffer ([#92](https://github.com/ankorstore/yokai/issues/92)) ([6e0c995](https://github.com/ankorstore/yokai/commit/6e0c9957095d3bc8d3ca77b610e472d4a0f238f0))
+
 ## 1.0.0 (2024-01-10)