+++ trace/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0](https://github.com/ankorstore/yokai/compare/trace/v1.0.0...trace/v1.1.0) (2024-02-23)
+
+
+### Features
+
+* **trace:** Added possibility to dump spans from test processor ([#96](https://github.com/ankorstore/yokai/issues/96)) ([a610572](https://github.com/ankorstore/yokai/commit/a6105722d71c7a5ace9dbde0b33062e44f8f3ada))
+
 ## 1.0.0 (2024-01-10)